[feature] - latLong module
