+++ client/src/actions/index.js
@@ -15,7 +15,6 @@ export function answerQuestion(answer, props, answersArr) {
 
 
 export function postQuestionListInput(input) {
-	console.log('we are in postQuestionListInput', input)
 	return dispatch => {
 		axios.get('https://jsonplaceholder.typicode.com/posts')
 		.then((data) => {
@@ -42,6 +41,34 @@ export function updateUserInfo(username, lastName, firstName, email) {
       email
     }
   }
+}
+
+export function getLatLng(){
+	return dispatch => {
+	  function getLocation() {
+	      if (navigator.geolocation) {
+	          navigator.geolocation.getCurrentPosition(showLocation);
+	      } else {
+	         console.log( "Geolocation is not supported by this browser.");
+	      }
+	    }
+	    function showLocation(positionA){
+	      console.log("Latitude: " + positionA.coords.latitude + " Longitude: ", + positionA.coords.longitude); 
+	      lat = positionA.coords.latitude;
+	      long = positionA.coords.longitude;
+	    }
+	    getLocation()
+	    .then((position) => {
+	    	console.log('received data inside getLatLng', position)
+	    	dispatch({
+	    		type: 'COORDINATES',
+	    		payload: position
+	    	})
+	    })
+	    .catch(function(reason){
+	      console.log('Handle rejected promise ('+reason+') here.');
+	    })
+	}
 }
 
 export function sendInputToState(props){
